Move selectable_items from utils to new header file
The `selectable_items` template is not really a utility but a piece of
functionality. On top of that, it is (currently) only used in
`cmoh::accessor_bundle`.
